Add support for partially initializing arrays, closes #504
